Remove duplicate check for libc++.
Some compilers accept -stdlib=libc++ in the compilation phase even if libc++ is not present on the system. In such cases, the check for CXXFLAGS passes but that of LDFLAGS fails. Since CXX is invoked with both CXXFLAGS and LDFLAGS, -stdlib=libc++ continues to be passed on the command line and all subsequent link operations fail.